+++ src/R3/SCR31D.C
@@ -566,7 +566,7 @@ void scroll() { /* Line 367, Address: 0x1027400 */
     }
 
     wD5 -= wD1; /* Line 568, Address: 0x1027c50 */
-    pHScrollWork += lD0.w.l / 2U; /* Line 569, Address: 0x1027c58 */
+    pHScrollWork += (unsigned int)lD0.w.l / 2; /* Line 569, Address: 0x1027c58 */
     waterdirec.w += 64; /* Line 570, Address: 0x1027c7c */
 
     zonescrsetsub0(&pHScrollBuff, (unsigned short**)&pHScrollWork, z31d_kawatbl, awasintbl, wD1, (unsigned short*)&lD2.w.l, &wD4); /* Line 572, Address: 0x1027c90 */
